Put the test case inside the base class
